Tables: move DataTable views into a View namespace, update references
