+++ src/Tables/Columns/Column.php
@@ -111,17 +111,6 @@ class Column
         return in_array($context, $this->contexts);
     }
 
-    // public function responsive($breakpoint)
-    // {
-    //     switch ($breakpoint) {
-    //         case 'sm': $this->addClass('hidden sm:table-cell'); break;
-    //         case 'md': $this->addClass('hidden md:table-cell'); break;
-    //         case 'lg': $this->addClass('hidden lg:table-cell'); break;
-    //     }
-
-    //     return $this;
-    // }
-
     /**
      * Sets the column description, often displayed as smaller text below the label.
      *
