+++ src/Tables/View/DataTableView.php
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace Gibbon\Tables\Renderer;
+namespace Gibbon\Tables\View;
 
 use Gibbon\View\View;
 use Gibbon\Domain\DataSet;
